+++ src/lib/control.ts
@@ -6,7 +6,7 @@ export const showSection = (section: string, resourceType: ResourceType): boolea
    *
    * Collections for example only hold other records and so don't have downloads.
    */
-  const notForDatasets: string[] = ['collections', 'series', 'scale', 'physical_size']
+  const notForDatasets: string[] = ['series', 'scale', 'physical_size']
   const notForProducts: string[] = ['services']
   const notForCollections: string[] = [
     'access',